Make test self containing
copy the model file from test/test2/case1.mod wich is based on model.9.
this way the test is self contained and independant of future changes to
src/model.9.FIFO-AverageLoad.mod. The down side is that future bugfixes
to model.9 should be ported, i.e. cherrypicked to this file. Maintaining
will thus be a bitch.

I prefer the selfcontainedness of the test which means the test will
always work over the possibility that I would have port changes form
src/model.9 to case1.mod